Rename RatingScale to RatingScaleLegacy
