+++ draft-packages/rating-scale/KaizenDraft/RatingScaleLegacy/helpers/determineSelectionFromKeyPress.ts
