+++ draft-packages/rating-scale/KaizenDraft/RatingScaleLegacy/styles.module.scss
