+++ draft-packages/rating-scale/KaizenDraft/RatingScaleLegacy/types.ts
