+++ draft-packages/rating-scale/index.ts
@@ -1,1 +1,6 @@
-export * from "./KaizenDraft/RatingScale/RatingScale"
+export * from "./KaizenDraft/RatingScaleLegacy/RatingScaleLegacy"
+export {
+  Scale,
+  ScaleItem,
+  ScaleValueResponse,
+} from "./KaizenDraft/RatingScaleLegacy/types"
